+++ src/components/modals/Modals.js
@@ -1,6 +1,6 @@
 // src/components/Modals.js
 import AddProfileModal from './AddProfileModal';
 import EditPinInputModal from './EditPinInputModal';
-import SelectPinInputModal from './SelectPinInputModal';
+import SelectPinInputModal from '../SelectPinInputModal';
 
 export { AddProfileModal, EditPinInputModal, SelectPinInputModal };
