+++ src/components/modals/StoryGeneratorModal.js
