+++ src/components/modals/VoiceInputModal.js
@@ -140,7 +140,7 @@ const VoiceInputModal = ({
                 </View>
                 <View style={styles.imageContainer}>
                   <Image
-                    source={require('../../assets/images/Wave.png')}
+                    source={require('../../../assets/images/Wave.png')}
                     style={styles.image}
                   />
                 </View>
@@ -153,7 +153,7 @@ const VoiceInputModal = ({
                     onPress={startRecording}
                     style={styles.roundButton}>
                     <Image
-                      source={require('../../assets/images/voice.png')}
+                      source={require('../../../assets/images/voice.png')}
                       style={styles.voiceImage}
                     />
                   </TouchableOpacity>
